requirements.txt
Django==1.8.1
Markdown==2.6.2
django-filter==0.9.2
djangorestframework==3.1.1
gunicorn==19.3.0
